+++ cmake/modules/Bridge.cmake
@@ -13,6 +13,7 @@ target_sources(kwik_bridge
             modules/bridge/event_adapter.cppm
             modules/bridge/js_table_data_source.cppm
             modules/bridge/js_lazy_list_source.cppm
+            modules/bridge/element_spec.cppm
     
     PRIVATE
         src/bridge/props_parser.cpp
@@ -24,6 +25,7 @@ target_sources(kwik_bridge
         src/bridge/event_adapter.cpp
         src/bridge/js_table_data_source.cpp
         src/bridge/js_lazy_list_source.cpp
+        src/bridge/element_spec.cpp
 )
 target_link_libraries(kwik_bridge
     PRIVATE